Create oddeven.cpp
In summary, this program reads an integer from the user and checks whether it is even or odd. It then prints "Even" if the number is even and "Odd" if the number is odd.

Explanation:
Include Libraries:

#include<iostream>: This library is used for input and output operations.

Namespace:

using namespace std;: This line allows us to use standard library names without the std:: prefix.

Main Function:

int main(): This is the main function where the execution of the program begins.

Input an Integer:

int n; cin >> n;: This line reads an integer n from the user.

Check if the Number is Even or Odd:

if(n % 2 == 0): This condition checks if n is divisible by 2 (i.e., if n is even).

cout << "Even" << endl;: If the condition is true, it prints "Even".

else: If the condition is false (i.e., if n is not divisible by 2), it executes the following block.

cout << "Odd" << endl;: It prints "Odd"

Return Statement:

return 0;: This indicates that the program ended successfully.