9.2DArray
Spiral Search
Spiral Search is a technique used to traverse a 2D matrix in a spiral order. This means starting from the top-left corner and moving right, then down, then left, and then up, repeating this pattern until all elements are visited

Transpose
Transpose of a matrix is an operation that flips a matrix over its diagonal, switching the row and column indices of the matrix,

Sort Search involves sorting an array and then performing a search operation, typically binary search, to find an element efficiently.

Summary:
Spiral Search: Traverses a 2D matrix in a spiral order.
Transpose: Flips a matrix over its diagonal.
Sort Search: Sorts an array and performs a binary search to find an element.